% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{madansky.fun}
\alias{madansky.fun}
\title{Lagrange multiplier in Madansky's method}
\usage{
madansky.fun(lam, s, n, alpha)
}
\arguments{
\item{lam}{The value of the Lagrange multiplier}

\item{s}{Vector of successes.}

\item{n}{Vector of sample sizes.}

\item{alpha}{The significance level; to calculate a 100(1-\eqn{\alpha})\% lower confidence bound.}
}
\description{
This function is called in the \code{madansky()} function to solve for the Lagrange multipliers.
}
