% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mann_grubbs}
\alias{mann_grubbs}
\title{Mann and Grubb's method}
\usage{
mann_grubbs(s, n, alpha, ...)
}
\arguments{
\item{s}{Vector of successes.}

\item{n}{Vector of sample sizes.}

\item{alpha}{The significance level; to calculate a 100(1-\eqn{\alpha})\% lower confidence bound.}

\item{...}{Additional arguments to be ignored.}
}
\value{
The 100(1-\eqn{\alpha})\% lower confidence bound.
}
\description{
Calculate a binomial series lower confidence bound using Mann and Grubb's (1974) method.
}
\examples{
mann_grubbs(s=c(35, 97, 59), n=c(35, 100, 60), alpha=.10)
}
