% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mann_grubbs_sum}
\alias{mann_grubbs_sum}
\title{Function to calculate the restricted sum in the Mann-Grubbs method.}
\usage{
mann_grubbs_sum(s, n)
}
\arguments{
\item{s}{Vector of successes.}

\item{n}{Vector of sample sizes.}
}
\value{
The restricted sum.
}
\description{
Calculate the restricted sum in the Mann-Grubbs method.
}
