% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mr.fun}
\alias{mr.fun}
\title{Function of \eqn{\beta} in the Myhre-Rennie 2 method}
\usage{
mr.fun(beta, s, n)
}
\arguments{
\item{beta}{The value of \eqn{\beta}.}

\item{s}{Vector of successes.}

\item{n}{Vector of sample sizes.}
}
\description{
This function is called in \code{myhre_rennie2()} function to solve for the \eqn{\beta} value.
}
