% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm.random}
\alias{pm.random}
\title{Matrix of \emph{p}-vector combinations sampled randomly.}
\usage{
pm.random(Rs.int, m, number)
}
\arguments{
\item{Rs.int}{Interval (or single number) of total system reliability.}

\item{m}{Number of components.}

\item{number}{The number of random samples to draw.}
}
\description{
Randomly sample to build a matrix of \emph{p}-vector combinations (component reliabilities) which lie in the specified interval of system reliability. Rows correspond to \emph{p}-vectors and columns correspond to components.
}
\examples{
pm.random(Rs.int=c(.9, .95), m=3, number=100)
}
