% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rmse.LCB}
\alias{rmse.LCB}
\title{Root Mean Square Error}
\usage{
rmse.LCB(LCB, R)
}
\arguments{
\item{LCB}{Vector of LCB's.}

\item{R}{The true system reliability .}
}
\value{
The root mean squared error of the LCB's from the true system reliability.
}
\description{
Calculate the root mean squared errors of the LCB's from the true system reliability. A measure of spread.
}
