% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{launch_app}
\alias{launch_app}
\title{Launch Shiny App}
\usage{
launch_app(MonteCarlo = 1000)
}
\arguments{
\item{MonteCarlo}{The number of Monte Carlo samples to take. E.g. In a Bayesian method, how many samples to take from a posterior distribution to estimate the lower \eqn{\alpha}-th quantile. The default value is 1000.}
}
\description{
Launches an instance of an R Shiny App, which runs locally on the user's computer.
}
\details{
If the "Download Histograms" button does not work, it can be fixed by launching the Shiny App on your local browser. This can be done by clicking on "Open in Browser" located at the top of your Shiny App. This seems to be an issue with the Download Handler that Shiny uses.
}
