% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normal_approximation}
\alias{normal_approximation}
\title{Normal approximation method}
\usage{
normal_approximation(s, n, alpha, ...)
}
\arguments{
\item{s}{Vector of successes.}

\item{n}{Vector of sample sizes.}

\item{alpha}{The significance level; to calculate a 100(1-\eqn{\alpha})\% lower confidence bound.}

\item{...}{Additional arguments to be ignored.}
}
\value{
The 100(1-\eqn{\alpha})\% lower confidence bound.
}
\description{
Calculate a binomial series lower confidence bound using a normal approximation with MLE estimates.
}
\examples{
normal_approximation(s=c(35, 97, 59), n=c(35, 100, 60), alpha=.10)
}
